pyproject.toml
[tool.poetry]
name = "ai-email"
version = "0.1.0"
description = ""
authors = ["Scott Lindsey <scott@scott.to>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pandas = "^2.2.2"
scikit-learn = "^1.4.2"
openai = "^1.26.0"
python-dotenv = "^1.0.1"
mail-parser = "^3.15.0"
imapclient = "^3.0.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
